extract action bar out fo board with game store provider Fixes #19
