+++ apps/game-of-life-react/src/App.tsx
@@ -1,15 +1,38 @@
 import { Box, MantineProvider } from "@mantine/core";
 import { AppShell, Navbar, Header } from "@mantine/core";
-import { GameOfLifeBoard } from "./components/GameOfLifeBoard";
+import { ActionBar } from "./components/ActionBar";
+import {
+  GameOfLifeContext,
+  GameOfLifeProvider,
+} from "./providers/GameOfLifeProvider";
+import { useContext } from "react";
+import { GameBoard } from "./components/GameBoard";
 
 function App() {
+  return (
+    <GameOfLifeProvider>
+      <AppInner />
+    </GameOfLifeProvider>
+  );
+}
+function AppInner() {
+  const { isStarted, height, width, toggle, setWidth, setHeight, board } =
+    useContext(GameOfLifeContext);
+
   return (
     <MantineProvider withGlobalStyles withNormalizeCSS>
       <AppShell
         padding="md"
         navbar={
           <Navbar width={{ base: 300 }} height={"100%"} p="xs">
-            {/* Navbar content */}
+            <ActionBar
+              isOn={isStarted}
+              height={height}
+              width={width}
+              onToggle={toggle}
+              onWidthChange={setWidth}
+              onHeightChange={setHeight}
+            />
           </Navbar>
         }
         header={
@@ -27,7 +50,7 @@ function App() {
         })}
       >
         <Box sx={{ height: "100%" }}>
-          <GameOfLifeBoard />
+          <GameBoard board={board} />
         </Box>
       </AppShell>
     </MantineProvider>
